requirements.txt
absl-py==0.2.2
astor==0.6.2
bleach==1.5.0
click==6.7
Flask==1.0.2
Flask-WTF==0.14.2
gast==0.2.0
gevent==1.3.5
greenlet==0.4.14
grpcio==1.13.0
gunicorn==19.9.0
html5lib==0.9999999
itsdangerous==0.24
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
numpy==1.14.5
Pillow==5.1.0
protobuf==3.6.0
scikit-learn==0.19.1
scipy==1.1.0
setuptools==39.1.0
six==1.11.0
tensorboard==1.10.0
tensorflow==1.10.0
tensorflow-serving-api==1.10.0
termcolor==1.1.0
Werkzeug==0.14.1
WTForms==2.2.1
WTForms-JSON==0.3.3
